Implemented smooth scrolling/zooming

+++ interfaces/IRenderer.hpp
@@ -8,7 +8,7 @@ class IRenderer {
 
 public:
   virtual void Delay(size_t ms) noexcept = 0;
-  virtual void Render(const Coord &pos, double scale,
+  virtual void Render(const Coord &pos, float scale,
                       const std::vector<std::string> &Objects) = 0;
   virtual ~IRenderer() noexcept = default;
 };
